Implement parsing for system-agnostic characters
The DSL used is implicit in the parser implementation and was more or
less invented in an /ad hoc/ manner.  The syntax is intended to be
natural and lightweight, and there is an example in the ScalaDoc for the
`AgnosticParser`.  Eventually, of course, examples will also be added to
the README.